refactor: 그룹 이미지 예외 추가
