+++ backend/src/test/java/com/woowacourse/momo/storage/service/GroupImageServiceTest.java
@@ -25,12 +25,13 @@ import org.springframework.transaction.annotation.Transactional;
 import com.woowacourse.momo.auth.support.SHA256Encoder;
 import com.woowacourse.momo.category.domain.Category;
 import com.woowacourse.momo.global.exception.exception.MomoException;
+import com.woowacourse.momo.group.domain.Group;
+import com.woowacourse.momo.group.domain.GroupName;
+import com.woowacourse.momo.group.domain.GroupRepository;
+import com.woowacourse.momo.group.domain.calendar.Calendar;
 import com.woowacourse.momo.group.domain.calendar.Deadline;
 import com.woowacourse.momo.group.domain.calendar.Schedules;
-import com.woowacourse.momo.group.domain.group.Capacity;
-import com.woowacourse.momo.group.domain.group.Group;
-import com.woowacourse.momo.group.domain.group.GroupName;
-import com.woowacourse.momo.group.domain.group.GroupRepository;
+import com.woowacourse.momo.group.domain.participant.Capacity;
 import com.woowacourse.momo.member.domain.Member;
 import com.woowacourse.momo.member.domain.MemberRepository;
 import com.woowacourse.momo.member.domain.Password;
@@ -91,8 +92,13 @@ class GroupImageServiceTest {
     }
 
     private Group saveGroup(String name, Category category) {
-        return groupRepository.save(new Group(new GroupName(name), savedHost, category, new Capacity(3),
-                이틀후부터_5일동안.toDuration(), new Deadline(내일_23시_59분.toDateTime()),
-                new Schedules(List.of(이틀후_10시부터_12시까지.toSchedule())), "", ""));
+        Calendar calendar = new Calendar(
+                new Deadline(내일_23시_59분.toDateTime()), 이틀후부터_5일동안.toDuration(),
+                new Schedules(List.of(이틀후_10시부터_12시까지.toSchedule()))
+        );
+        Group group = new Group(
+                savedHost, new Capacity(3), calendar, new GroupName(name), category, "", ""
+        );
+        return groupRepository.save(group);
     }
 }
